+++ packages/runtime/src/interpreter/opcode-list.ts
@@ -24,21 +24,24 @@ import { RUNTIME_ERRORS } from "../errors/errors-list";
 import { RuntimeError } from "../errors/runtime-errors";
 import { compareArray } from "../lib/compare";
 import {
+	AccountParamGetField,
 	AcctParamQueryFields,
 	AcctParamQueryFieldsIndex,
 	ALGORAND_MAX_LOGS_COUNT,
 	ALGORAND_MAX_LOGS_LENGTH,
 	AppParamDefined,
 	AppParamDefinedIndex,
+	AppParamField,
 	AssetHoldingField,
 	AssetParamGetField,
-	AppParamField,
 	AssetParamMap,
 	AssetParamMapIndex,
+	Base64Encoding,
 	blockFieldIndex,
 	blockFieldTypes,
 	CurveTypeArgument,
 	CurveTypeEnum,
+	GlobalField,
 	GlobalFields,
 	GlobalFieldsIndex,
 	ITxArrFields,
@@ -57,13 +60,10 @@ import {
 	publicKeyLength,
 	TransactionTypeEnum,
 	TxArrFields,
+	TxFieldEnum,
 	TxnFieldsIndex,
 	vrfVerifyFieldTypes,
 	ZERO_ADDRESS,
-	TxFieldEnum,
-	Base64Encoding,
-	GlobalField,
-	AccountParamGetField
 } from "../lib/constants";
 import { addInnerTransaction, calculateInnerTxCredit, setInnerTxField } from "../lib/itxn";
 import { bigintSqrt } from "../lib/math";
@@ -1451,15 +1451,7 @@ export class Txn extends Op {
 		this.line = line;
 		this.idx = undefined;
 
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = TxnFieldsIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
-
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
 		this.assertTxFieldDefined(argument, interpreter.tealVersion, line);
 		if (TxArrFields[interpreter.tealVersion].has(argument)) {
 			// eg. txn Accounts 1
@@ -1589,14 +1581,7 @@ export class Txna extends Op {
 		super();
 		this.line = line;
 		assertOnlyDigits(args[1], line);
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = TxnFieldsIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
 		this.assertTxArrFieldDefined(argument, interpreter.tealVersion, line);
 
 		this.field = argument; // field
@@ -1653,14 +1638,7 @@ export class Gtxna extends Op {
 		assertOnlyDigits(args[0], line);
 		assertOnlyDigits(args[2], line);
 
-		let argument = args[1];
-		try {
-			const index = Number(args[1]);
-			argument = TxnFieldsIndex[index];
-			if (argument === undefined) argument = args[1];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[1], TxnFieldsIndex);
 		this.assertTxArrFieldDefined(argument, interpreter.tealVersion, line);
 
 		this.txIdx = Number(args[0]); // transaction group index
@@ -1890,14 +1868,7 @@ export class Global extends Op {
 	constructor(args: string[], line: number, interpreter: Interpreter) {
 		super();
 		assertLen(args.length, 1, line);
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = GlobalFieldsIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], GlobalFieldsIndex);
 		this.assertGlobalDefined(argument, interpreter.tealVersion, line);
 
 		this.field = argument; // global field
@@ -2389,11 +2360,11 @@ export class GetAssetHolding extends Op {
 		}
 		let value: StackElem;
 		switch (this.field) {
-      case "0":
+			case "0":
 			case AssetHoldingField.AssetBalance:
 				value = BigInt(assetInfo.amount);
 				break;
-      case "1":
+			case "1":
 			case AssetHoldingField.AssetFrozen:
 				value = assetInfo["is-frozen"] ? 1n : 0n;
 				break;
@@ -2428,14 +2399,7 @@ export class GetAssetDef extends Op {
 		this.line = line;
 		this.interpreter = interpreter;
 		assertLen(args.length, 1, line);
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = AssetParamMapIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], AssetParamMapIndex);
 		if (AssetParamMap[interpreter.tealVersion][argument] === undefined) {
 			throw new RuntimeError(RUNTIME_ERRORS.TEAL.UNKNOWN_ASSET_FIELD, {
 				field: args[0],
@@ -2480,7 +2444,9 @@ export class GetAssetDef extends Op {
 					break;
 				default:
 					def = AssetDefinition[s] as string;
-					value = isValidAddress(def) ? decodeAddress(def).publicKey : parsing.stringToBytes(def);
+					value = isValidAddress(def)
+						? decodeAddress(def).publicKey
+						: parsing.stringToBytes(def);
 					break;
 			}
 
@@ -4109,8 +4075,8 @@ export class EcdsaPkRecover extends Op {
 		const signatureC = this.assertBytes(stack.pop(), this.line);
 		const recoverId = this.assertBigInt(stack.pop(), this.line);
 		const data = this.assertBytes(stack.pop(), this.line);
-    
-    const ec = new EC(this.curveType);
+
+		const ec = new EC(this.curveType);
 		const signature = { r: signatureC, s: signatureD };
 		const pubKey = ec.recoverPubKey(data, signature, Number(recoverId));
 		const x = pubKey.getX();
@@ -4316,14 +4282,7 @@ export class ITxnField extends Op {
 		super();
 		this.line = line;
 
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = TxnFieldsIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
 		this.assertTxFieldDefined(argument, interpreter.tealVersion, line);
 		assertLen(args.length, 1, line);
 		this.field = argument; // field
@@ -4449,10 +4408,10 @@ export class ITxnSubmit extends Op {
 			const signedTransactions: algosdk.SignedTransaction[] = execParams.map((txnParam) =>
 				types.isExecParams(txnParam)
 					? {
-						sig: Buffer.alloc(5),
-						sgnr: Buffer.from(algosdk.decodeAddress(contractAddress).publicKey),
-						txn: webTx.mkTransaction(txnParam, mockSuggestedParams(txnParam.payFlags, 1)),
-					}
+							sig: Buffer.alloc(5),
+							sgnr: Buffer.from(algosdk.decodeAddress(contractAddress).publicKey),
+							txn: webTx.mkTransaction(txnParam, mockSuggestedParams(txnParam.payFlags, 1)),
+					  }
 					: txnParam
 			);
 			this.interpreter.runtime.ctx.processTransactions(signedTransactions);
@@ -4497,14 +4456,7 @@ export class ITxn extends Op {
 		this.line = line;
 		this.idx = -1;
 
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = TxnFieldsIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
 		this.assertITxFieldDefined(argument, interpreter.tealVersion, line);
 		if (
 			TxArrFields[interpreter.tealVersion].has(argument) ||
@@ -4906,14 +4858,7 @@ export class AcctParamsGet extends Op {
 		this.line = line;
 		this.interpreter = interpreter;
 		assertLen(args.length, 1, line);
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = AcctParamQueryFieldsIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], AcctParamQueryFieldsIndex);
 		if (
 			!AcctParamQueryFields[argument] ||
 			AcctParamQueryFields[argument].version > interpreter.tealVersion
@@ -5108,7 +5053,7 @@ export class Base64Decode extends Op {
 				this.encoding = "base64url";
 				break;
 			}
-      case "1":
+			case "1":
 			case Base64Encoding.StdEncoding: {
 				this.encoding = "base64";
 				break;
@@ -5432,14 +5377,7 @@ export class Block extends Op {
 	constructor(args: string[], line: number, interpreter: Interpreter) {
 		super();
 		assertLen(args.length, 1, line);
-		let argument = args[0];
-		try {
-			const index = Number(args[0]);
-			argument = blockFieldIndex[index];
-			if (argument === undefined) argument = args[0];
-		} catch {
-			//Bind value index to asset map
-		}
+		const argument = this.assertFieldIndex(args[0], blockFieldIndex);
 		if (argument === blockFieldTypes.BlkSeed || argument === blockFieldTypes.BlkTimestamp) {
 			this.field = argument;
 		} else {